Rename appReducer to mainReducer. Singular file names.
