+++ src/reducers/catalogReducer.js
